Lots of bug fixes for filters, pipelines, other classes and pipeline notifications. (#81)
+ DataArray
  - Add early return to copyTuple if the source and destination are the source
  - Add getTypeName() which returns more information in the string about the type 'DataArray<uint8_t>'

+ DataStore
  - Remove commennted code

+ IDataArray
  - Added virtual getTypeName() method

+ NeighborList()
  - Add public default constructor
  - Add some DOxygen doc skeleton

+ DataArrayFactory
  - Added method to get the 'typeName' from the actual DataArray<T>
  - Supports reading .dream3d (v8) files because of the change to getTypeName()

+ ArrayThresholdsParmaeter
  - Refactored a few variables
  - Fixed an error message

+ Pipeline
  - Added std::cout for each filter. Assists debugging (probably not needed long term)

+ DataArrayUtilities
  - Updated ReplaceValue() with additional template parameters in support of redesigned Threshold filter

+ DataGroupUtilities
  - New header class for DataGroup related operations
  - Added RemoveInactiveObjects() in support of RemoveMinimumSizeFilter

+ H5DataFactoryManager
  - Added new factories for each of the DataArray Types

+ CalculateFeatureSizes
  - Set tuple dimensions to {1} during preflight
  - Ensure created DataArrays have the proper Tuple Dimensions

+ ConditionalSetValue
  - Ensure we are checking DataArrays are the types expected by the filter.
  - Create constants for error codes

+ CreateDataArray
  - Added DynamicaTableData parameter to get the Tuple Dimensions from the user.
 - Create Data Array should have a starting 1 row x 1 column start.

+ FindDifferencesMap
  - ClangTidy misc fixes
  - Ensure output DataArray has the proper TupleShape

+ MultiThresholdObjects
  - Rewrote the threshold classes for huge performance gain. Now as fast as SIMPL based codes
  - These updates need further unit testing

+ RawBinaryReaderFilter
  - Added some help text for the Tuple Dims parameter

+ RemoveMinimunSizeFeaturesFilter
  - Added preflight checks for valid DataArrays
  - Finished implementing the final part of the filter where the feature data arrays are updated

+ ScalarSegmentFeatures | SegmentFeatures
  - Moved most of the code to an 'Algorithms' class
  - ScalarSegmentFeaturesFilter no longer subclasses from SegmentFeaturesFilter
  - Redesigned the ScalarSegmentFeatures algorithm for large performance gain

+ Adding Pipeline and Filter messaging using nod signals.

These signals are specific to running pipelines and not mean to replace the general notification system

+ Add Queued status to RunState

+ Remove Pipeline::Status & replace w/ RunState & FaultState data members

+ Skip disabled filters when executing Pipeline

+ Send filter fault state message signal from PipelineFilter object instead of Pipeline

+ Add missing filter fault state signal when filter fails
+ Add RunState and FaultState updates to preflighting
+ Skip preflighting filter when it is disabled
+ Add missing filter fault detail messages during preflight

+ ConditionalSetValue uses IDataArray->getDataType() instead of string comparisons
+ Move RunState and FaultState out of AbstractPipelineNode.
+ Cleared error message in DeleteDataAction

+ Export DREAM3D File
  - Remove error when final file path does not exist as the underlying code will create the path as needed.

+ MultiThresholdObjects: Check for empty datapaths and return if empty.
  - The rest of preflightImpl cannot do anything with empty datapaths.

+ Create specific ProgressMessage struct.
+ use complex types in GetTypeName

+++ src/Plugins/ComplexCore/src/ComplexCore/Filters/IterativeClosestPointFilter.cpp
@@ -20,10 +20,10 @@ namespace complex
 {
 namespace
 {
-constexpr int64 k_MissingMovingVertex = -4500;
-constexpr int64 k_MissingTargetVertex = -4501;
-constexpr int64 k_BadNumIterations = -4502;
-constexpr int64 k_MissingVertices = -4503;
+constexpr int32 k_MissingMovingVertex = -4500;
+constexpr int32 k_MissingTargetVertex = -4501;
+constexpr int32 k_BadNumIterations = -4502;
+constexpr int32 k_MissingVertices = -4503;
 
 template <typename Derived>
 struct VertexGeomAdaptor
